+++ src/test/java/com/selenium101/app/AppTest.java
@@ -1,6 +1,7 @@
 package com.selenium101.app;
 
-import com.selenium101.app.pageObject.GooglePage;
+import com.selenium101.app.pageObject.BasePage;
+import com.selenium101.app.pageObject.GoogleSearchPage;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -11,7 +12,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 
 public class AppTest {
     private static WebDriver driver;
-    private GooglePage GooglePage;
+    private GoogleSearchPage GooglePage;
 
     @BeforeClass
     public static void beforeClass() throws Exception {
@@ -21,7 +22,8 @@ public class AppTest {
 
     @Before
     public void setup() {
-        GooglePage = new GooglePage(driver);
+        GooglePage = new GoogleSearchPage(driver);
+
         driver.get("http://google.com");
     }
